Revert custom audio policy configuration.
The gts2-common default audio policy configuration turned out to be
the best match for the majority of users, according to the XDA thread poll at
https://xdaforums.com/t/unofficial-rom-alpha-lineageos-20-for-sm-t715-october-1-2023.4606053/page-3
So reverting back to the configuration that has
- Bluetooth audio through bluetooth loudspeaker
- Telephony audio using a wired headset
working.
The downside is that telephony audio with built-in speaker and microphone
is not working with this configuration.
If someone wants that config, it is still available at
https://github.com/metterschlg/android_device_samsung_gts28ltexx/blob/lineage-20/configs/audio/audio_policy_configuration.xml
and can be applied using the method described at
https://xdaforums.com/t/unofficial-rom-alpha-lineageos-20-for-sm-t715-october-1-2023.4606053/post-88870275

Signed-off-by: metterschlg <106471001+metterschlg@users.noreply.github.com>