LineageOS 21 bringup plus long-standing "wifionly" bug fix.
For LOS21 we need to move some legacy HIDL manifests to a separate file.
Also take care to build on top of the gts2-common manifests.

As we got the tablet marked as wifionly, then added telephony attributes,
this strategy got broken with recent LineageOS changes for tablet only
devices.
The non-wifionly tablet configuration includes the telephony features,
so proceed with that one.

Signed-off-by: metterschlg <106471001+metterschlg@users.noreply.github.com>